+++ proofs/stream.php
@@ -5,10 +5,13 @@ use Innmind\IO\{
     Stream\Size,
     Stream\Size\Unit,
 };
-use Innmind\BlackBox\Set;
+use Innmind\BlackBox\{
+    Set,
+    Prove,
+};
 
-return static function() {
-    yield test(
+return static function(Prove $prove) {
+    yield $prove->test(
         'Empty stream size',
         static function($assert) {
             $model = Size::of(0);
@@ -18,13 +21,13 @@ return static function() {
         },
     );
 
-    yield proof(
-        'Stream sizes',
-        given(
+    yield $prove
+        ->proof('Stream sizes')
+        ->given(
             Set::integers()->between(1, 999),
             Set::of(...Unit::cases()),
-        ),
-        static function($assert, $size, $unit) {
+        )
+        ->test(static function($assert, $size, $unit) {
             $model = $unit->of($size);
             $extension = match ($unit) {
                 Unit::bytes => 'B',
@@ -37,21 +40,20 @@ return static function() {
 
             $assert->same($unit, $model->unit());
             $assert->same("$size$extension", $model->toString());
-        },
-    );
+        });
 
-    yield proof(
-        'Stream::lessThan()',
-        given(
+    yield $prove
+        ->proof('Stream::lessThan()')
+        ->given(
             Set::integers()->above(0),
             Set::integers()->above(1),
-        )->filter(static fn($a, $b) => \is_int($a + $b)),
-        static function($assert, $size, $additionnal) {
+        )
+        ->filter(static fn($a, $b) => \is_int($a + $b))
+        ->test(static function($assert, $size, $additionnal) {
             $model = Size::of($size);
 
             $assert->false($model->lessThan($model));
             $assert->true($model->lessThan(Size::of($size + $additionnal)));
             $assert->false(Size::of($size + $additionnal)->lessThan($model));
-        },
-    );
+        });
 };